+++ src/bgp/mod.rs
@@ -1,7 +1,9 @@
 mod announcer;
 mod gobgp;
 mod mock;
+mod proto;
 
 pub use announcer::*;
 pub use gobgp::*;
 pub use mock::*;
+pub(crate) use proto::apipb;