+++ src/bgp/proto.rs
@@ -1,0 +1,3 @@
+pub mod apipb {
+    tonic::include_proto!("apipb");
+}